+++ src/main/java/br/com/collegesmaster/business/impl/DisciplineBusinessImpl.java
@@ -1,13 +1,19 @@
 package br.com.collegesmaster.business.impl;
 
+import static br.com.collegesmaster.rest.util.RestUtils.buildPredicatesFromRequest;
+import static java.lang.Boolean.FALSE;
+import static java.lang.Boolean.TRUE;
+import static javax.ejb.TransactionAttributeType.REQUIRED;
 import static javax.ejb.TransactionManagementType.CONTAINER;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Map;
 
 import javax.annotation.security.PermitAll;
 import javax.annotation.security.RolesAllowed;
 import javax.ejb.Stateless;
+import javax.ejb.TransactionAttribute;
 import javax.ejb.TransactionManagement;
 import javax.inject.Inject;
 import javax.persistence.EntityManager;
@@ -17,6 +23,7 @@ import javax.persistence.criteria.CriteriaQuery;
 import javax.persistence.criteria.Predicate;
 import javax.persistence.criteria.Root;
 import javax.persistence.criteria.Selection;
+import javax.ws.rs.core.UriInfo;
 
 import org.jboss.ejb3.annotation.SecurityDomain;
 
@@ -38,27 +45,53 @@ public class DisciplineBusinessImpl implements DisciplineBusiness {
 	@Inject
 	protected CriteriaBuilder cb;
 	
+	@TransactionAttribute(REQUIRED)
 	@Override
-	public void create(DisciplineImpl discipline) {
-		em.persist(discipline);
+	public Boolean create(DisciplineImpl discipline) {
+		if(discipline != null && discipline.getId() == null && discipline.getVersion() == null) {
+			em.persist(discipline);
+			return TRUE;
+		} else {
+			LOGGER.warn("Entity not persisted, invalid arguments");
+			return FALSE;			
+		}
 	}
 
+	@TransactionAttribute(REQUIRED)
 	@Override
 	public DisciplineImpl update(DisciplineImpl discipline) {
-		return em.merge(discipline);
+		if(discipline != null && discipline.getId() != null && discipline.getVersion() != null) {
+			return em.merge(discipline);
+		} else {
+			LOGGER.warn("Entity not persisted, invalid arguments");
+			return null;
+		}
 	}
 
+	@TransactionAttribute(REQUIRED)
 	@Override
-	public void remove(DisciplineImpl discipline) {
-		em.remove(discipline);
+	public Boolean remove(DisciplineImpl discipline) {
+		if(discipline != null && discipline.getId() != null && discipline.getVersion() != null) {
+			em.remove(discipline);				
+			return TRUE;
+		} else {
+			LOGGER.warn("Entity not removed, invalid arguments");
+			return FALSE;
+		}
 	}
-
-	@Override
+	
 	@PermitAll
+	@Override
 	public DisciplineImpl findById(Integer id) {
-		return em.find(DisciplineImpl.class, id);
+		if(id != null) {
+			return em.find(DisciplineImpl.class, id);			
+		} else {
+			LOGGER.warn("Cannot find entity, invalid arguments");
+			return null;
+		}
 	}
-
+	
+	@TransactionAttribute(REQUIRED)
 	@Override
 	public List<DisciplineImpl> findAll() {
 		
@@ -68,6 +101,28 @@ public class DisciplineBusinessImpl implements DisciplineBusiness {
 		final List<DisciplineImpl> result = typedQuery.getResultList(); 
 		
 		return result;
+	}
+	
+	@PermitAll
+	@TransactionAttribute(REQUIRED)
+	@Override
+	public List<DisciplineImpl> findAll(final UriInfo requestInfo) {
+		
+		final Map<String, Object> equalsPredicate = buildPredicatesFromRequest(requestInfo, DisciplineImpl.class);
+		
+		final CriteriaQuery<DisciplineImpl> criteriaQuery = cb.createQuery(DisciplineImpl.class);		
+		final Root<DisciplineImpl> instituteRoot = criteriaQuery.from(DisciplineImpl.class);
+		
+		final List<Predicate> predicates = new ArrayList<>();
+		equalsPredicate.forEach((key, value) -> {
+			predicates.add(cb.equal(instituteRoot.get(key), value));
+		});
+
+		criteriaQuery.where(predicates.toArray(new Predicate[0]));
+
+		final TypedQuery<DisciplineImpl> typedQuery = em.createQuery(criteriaQuery);		
+		
+		return typedQuery.getResultList();
 	}
 	
 	@PermitAll
