+++ src/main/java/br/com/collegesmaster/business/impl/UserBusinessImpl.java
@@ -1,11 +1,14 @@
 package br.com.collegesmaster.business.impl;
 
+import static br.com.collegesmaster.rest.util.RestUtils.buildPredicatesFromRequest;
 import static java.lang.Boolean.FALSE;
 import static java.lang.Boolean.TRUE;
 import static javax.ejb.TransactionAttributeType.REQUIRED;
 import static javax.ejb.TransactionManagementType.CONTAINER;
 
+import java.util.ArrayList;
 import java.util.List;
+import java.util.Map;
 
 import javax.annotation.security.PermitAll;
 import javax.annotation.security.RolesAllowed;
@@ -22,6 +25,7 @@ import javax.persistence.criteria.CriteriaQuery;
 import javax.persistence.criteria.Predicate;
 import javax.persistence.criteria.Root;
 import javax.persistence.criteria.Subquery;
+import javax.ws.rs.core.UriInfo;
 
 import org.jboss.ejb3.annotation.SecurityDomain;
 import org.jboss.logging.Logger;
@@ -55,28 +59,50 @@ public class UserBusinessImpl implements UserBusiness {
 
 	@PermitAll
 	@Override
-	public void create(final UserImpl user) {
-		em.persist(user);
+	public Boolean create(final UserImpl user) {
+		if(user != null && user.getId() == null && user.getVersion() == null) {
+			em.persist(user);
+			return TRUE;
+		} else {
+			LOGGER.warn("Entity not persisted, invalid arguments");
+			return FALSE;			
+		}
 	}
 
 	@RolesAllowed({ "STUDENT", "PROFESSOR", "ADMINISTATOR" })
 	@Override
 	public UserImpl update(final UserImpl user) {
-		final UserImpl updatedUser = em.merge(user);
-		userUpdateEvent.fire(updatedUser);
-		return updatedUser;
+		if(user != null && user.getId() != null && user.getVersion() != null) {
+			final UserImpl updatedUser = em.merge(user);
+			userUpdateEvent.fire(updatedUser);
+			return updatedUser;
+		} else {
+			LOGGER.warn("Entity not persisted, invalid arguments");
+			return null;
+		}
 	}
 
 	@RolesAllowed({ "ADMINISTRATOR" })
 	@Override
-	public void remove(final UserImpl user) {
-		em.remove(user);
+	public Boolean remove(final UserImpl user) {
+		if(user != null && user.getId() != null && user.getVersion() != null) {
+			em.remove(user);				
+			return TRUE;
+		} else {
+			LOGGER.warn("Entity not removed, invalid arguments");
+			return FALSE;
+		}
 	}
 
 	@RolesAllowed({ "STUDENT", "PROFESSOR", "ADMINISTATOR" })
 	@Override
 	public UserImpl findById(final Integer id) {
-		return em.find(UserImpl.class, id);
+		if(id != null) {
+			return em.find(UserImpl.class, id);			
+		} else {
+			LOGGER.warn("Cannot find entity, invalid arguments");
+			return null;
+		}
 	}
 
 	@RolesAllowed({ "ADMINISTRATOR" })
@@ -89,7 +115,29 @@ public class UserBusinessImpl implements UserBusiness {
 
 		return result;
 	}
+	
+	@RolesAllowed({"ADMINISTRATOR"})
+	@TransactionAttribute(REQUIRED)
+	@Override
+	public List<UserImpl> findAll(final UriInfo requestInfo) {
+		
+		final Map<String, Object> equalsPredicate = buildPredicatesFromRequest(requestInfo, UserImpl.class);
+		
+		final CriteriaQuery<UserImpl> criteriaQuery = cb.createQuery(UserImpl.class);		
+		final Root<UserImpl> instituteRoot = criteriaQuery.from(UserImpl.class);
+		
+		final List<Predicate> predicates = new ArrayList<>();
+		equalsPredicate.forEach((key, value) -> {
+			predicates.add(cb.equal(instituteRoot.get(key), value));
+		});
 
+		criteriaQuery.where(predicates.toArray(new Predicate[0]));
+
+		final TypedQuery<UserImpl> typedQuery = em.createQuery(criteriaQuery);		
+		
+		return typedQuery.getResultList();
+	}
+	
 	@PermitAll
 	@TransactionAttribute(REQUIRED)
 	@Override
